% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_utils.R
\name{ggplotMinAethetics}
\alias{ggplotMinAethetics}
\title{ggplot minimal aesthetics}
\usage{
ggplotMinAethetics(
  ggplt,
  width = NULL,
  height = NULL,
  title = NULL,
  expand = c(0, 0),
  plot.title = element_text(size = 16),
  strip.text = element_text(size = 11),
  axis.text = element_text(size = 14),
  axis.title = element_text(size = 16),
  xlabel = NULL,
  ylabel = NULL,
  xlim = NULL,
  ylim = NULL,
  x.text.angle = NULL,
  legend.position = "right",
  legend.title = element_text(size = 16),
  legend.text = element_text(size = 16),
  legend.title.align = 0.5,
  legend.size = NULL,
  legend.size.unit = "in",
  scale_color = NULL,
  scale_fill = NULL,
  scale_fill_manual.values = NULL,
  scale_color_manual.values = NULL,
  scale_color_manual.breaks = waiver()
)
}
\arguments{
\item{ggplt}{ggplot object on minimal aesthetics is performed}

\item{width}{Set plot width}

\item{height}{Set plot height}

\item{expand}{Move the plot origin (default c(0,0)). The ggplot default value is restored with expand=waiver()}

\item{plot.title}{plot title (text appearance) (\code{\link[=element_text]{element_text()}}; inherits
from \code{title}) left-aligned by default}

\item{strip.text}{facet labels (\code{\link[=element_text]{element_text()}};
inherits from  \code{text}). Horizontal facet labels (\code{strip.text.x}) & vertical
facet labels (\code{strip.text.y}) inherit from \code{strip.text} or can be specified
separately}

\item{legend.position}{the position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector)}

\item{scale_color}{change line colors using the ggplot functions}

\item{scale_fill}{change fill colors using the ggplot functions}
}
\description{
This function renders high quality ggplots with minimal aesthetics.
}
\details{
List of theme parameters from https://github.com/tidyverse/ggplot2/blob/HEAD/R/theme.r
}
