% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_utils.R
\name{save_plot}
\alias{save_plot}
\title{Save ggplot to Multiple Formats}
\usage{
save_plot(
  plot,
  filename,
  width = NULL,
  height = NULL,
  file_formats = c("png", "svg")
)
}
\arguments{
\item{plot}{ggplot object to save}

\item{filename}{Character. Base filename (without extension)}

\item{width}{Numeric. Plot width (default: 3.7)}

\item{height}{Numeric. Plot height (default: 7.6)}

\item{file_formats}{Character vector. File formats to save}
}
\value{
Invisible NULL (called for side effects)
}
\description{
Helper function to save ggplot objects to multiple file formats
}
